[minor] abs
git-svn-id: https://subversion.renater.fr/wimsdev/trunk@8860 b48fb02d-6135-0410-b1dd-deb767c81da9

+++ wims/src/Misc/lceb/lceb_rec.c
@@ -101,7 +101,7 @@ int goal;
 void resultest(struct result * res)
 {
 	int tmp;
-	if ((tmp = res->val - goal) < min)
+	if ((tmp = (int)res->val - goal) < min)
 	{
 		min = tmp;
 		best = res;
